📚 Docs: Home Service icon 추가
