+++ src/constants/commonUiData.ts
@@ -57,46 +57,3 @@ export const HOME_OFFICE = [
     content: '10월 입주자 정기회의 일정 안내'
   }
 ];
-
-export const HOME_SERVICES = [
-  {
-    title: '오피스정보',
-    icon: ''
-  },
-  {
-    title: '불편접수',
-    icon: ''
-  },
-  {
-    title: '시설이용',
-    icon: ''
-  },
-  {
-    title: '차량관리',
-    icon: ''
-  },
-  {
-    title: '설문',
-    icon: ''
-  },
-  {
-    title: '구내식당',
-    icon: ''
-  },
-  {
-    title: '납부고지서',
-    icon: ''
-  },
-  {
-    title: '출입카드',
-    icon: ''
-  },
-  {
-    title: '시설안내',
-    icon: ''
-  },
-  {
-    title: '엘리베이터',
-    icon: ''
-  }
-];
