+++ src/constants/serviceMenus.ts
@@ -1,0 +1,53 @@
+import OfficeInfo from '@/assets/menus/ico_office_info.svg';
+import Inconvenience from '@/assets/menus/ico_inconvenience.svg';
+import UseGuide from '@/assets/menus/ico_use_guide.svg';
+import CarManage from '@/assets/menus/ico_car_manage.svg';
+import Restaurant from '@/assets/menus/ico_restaurant.svg';
+import Biil from '@/assets/menus/ico_bill.svg';
+import BuildingInfo from '@/assets/menus/ico_building_info.svg';
+import Elevator from '@/assets/menus/ico_elevator.svg';
+import Pass from '@/assets/menus/ico_pass.svg';
+import Survey from '@/assets/menus/ico_survey.svg';
+
+export const HOME_SERVICES = [
+  {
+    title: '오피스정보',
+    icon: OfficeInfo
+  },
+  {
+    title: '불편접수',
+    icon: Inconvenience
+  },
+  {
+    title: '시설이용',
+    icon: UseGuide
+  },
+  {
+    title: '차량관리',
+    icon: CarManage
+  },
+  {
+    title: '설문',
+    icon: Survey
+  },
+  {
+    title: '구내식당',
+    icon: Restaurant
+  },
+  {
+    title: '납부고지서',
+    icon: Biil
+  },
+  {
+    title: '출입카드',
+    icon: Pass
+  },
+  {
+    title: '시설안내',
+    icon: BuildingInfo
+  },
+  {
+    title: '엘리베이터',
+    icon: Elevator
+  }
+];
